test(tui): harden Windows terminal lifecycle
Refs #445
